requirements.txt
#Requirements for Learn Langchain Pinecone OpenAI
#pip install -r .\requirements.txt
openai
langchain
docx2txt
streamlit
chromadb                   
python-dotenv
tiktoken